savespot добавляю админку убрал лишнее
