+++ goods/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class PosobieConfig(AppConfig):
+class GoodsConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'posobie'
+    name = 'goods'
