+++ moscowliuda/settings.py
@@ -15,8 +15,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "authorization.apps.AuthorizationConfig",
     "moscowliuda_main.apps.MoscowliudaMainConfig",
-    "posobie.apps.PosobieConfig",
-    "project.apps.ProjectConfig",
+    "goods.apps.GoodsConfig",
 ]
 
 MIDDLEWARE = [
@@ -82,5 +81,3 @@ STATIC_URL = '/static/'
 MEDIA_URL = '/media/'
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 LOGIN_REDIRECT_URL = 'home'
-
-bot = "6800111134:AAGE54IqInIC68_nByP-HYTt40QbASzuzVU"
